+++ src/creatures/monsters/monster.h
@@ -297,7 +297,7 @@ class Monster final : public Creature {
 	void clearTargetList();
 	void clearFriendList();
 
-		BlockType_t blockHit(Creature* attacker, CombatType_t combatType, int64_t &damage, bool checkDefense = false, bool checkArmor = false, bool field = false) override;
+		BlockType_t blockHit(Creature* attacker, CombatType_t combatType, int32_t &damage, bool checkDefense = false, bool checkArmor = false, bool field = false) override;
 
 	static uint32_t monsterAutoID;
